move taxonomyLookup to forwardingRoute #1144
