+++ src/server/helpers/taxonomyLookup.js
@@ -5,16 +5,3 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
-
-import {
-  resolveJsonOrRejectWithError,
-  apiResourceUrl,
-  fetchWithAccessToken,
-} from '../../util/apiHelpers';
-
-const baseUrl = apiResourceUrl('/taxonomy/v1/url/mapping');
-
-export const taxonomyLookup = url => {
-  const k = `${baseUrl}?url=${url}`;
-  return fetchWithAccessToken(k).then(resolveJsonOrRejectWithError);
-};
